+++ src/main/java/insane96mcp/shieldsplus/mixin/PlayerMixin.java
@@ -1,5 +1,6 @@
 package insane96mcp.shieldsplus.mixin;
 
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import insane96mcp.shieldsplus.module.BaseFeature;
 import insane96mcp.shieldsplus.setup.SPEnchantments;
 import insane96mcp.shieldsplus.world.item.enchantment.FastRecoveryEnchantment;
@@ -11,9 +12,7 @@ import net.minecraft.world.level.Level;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Constant;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.ModifyConstant;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(Player.class)
@@ -25,13 +24,13 @@ public abstract class PlayerMixin extends LivingEntity {
         super(p_20966_, p_20967_);
     }
 
-    @ModifyConstant(constant = @Constant(floatValue = 3.0F), method = "hurtCurrentlyUsedShield")
-    private float blockingWindupTime(float minDamage) {
+    @ModifyExpressionValue(method = "hurtCurrentlyUsedShield", at = @At(value = "CONSTANT", args = "floatValue=3.0"))
+    private float shieldsPlus$blockingWindupTime(float minDamage) {
         return BaseFeature.minShieldHurtDamage.floatValue();
     }
 
     @Inject(method = "disableShield", at = @At("HEAD"), cancellable = true)
-    private void disableShield(boolean efficiencyAffected, CallbackInfo callbackInfo) {
+    private void shieldsPlus$disableShield(boolean efficiencyAffected, CallbackInfo callbackInfo) {
         if (BaseFeature.combatTestShieldDisabling()) {
             callbackInfo.cancel();
             int ticks = 32;